New base test case class BaseCaseRun
This base class contains pre-created test run and case runs.
